# TSDB: Shrink `memSeries` by using a single pointer and two bools for last histogram and float histogram values