--- a/tsdb/head_test.go
+++ b/tsdb/head_test.go
@@ -142,14 +142,14 @@ func BenchmarkCreateSeries(b *testing.B) {
 	}
 }
 
-func BenchmarkHeadAppender_Append_Commit_ExistingSeries(b *testing.B) {
+func BenchmarkHeadAppender_Append_Commit_ExistingSeries_FloatValues(b *testing.B) {
 	seriesCounts := []int{100, 1000, 10000}
 	series := genSeries(10000, 10, 0, 0)
 
 	for _, seriesCount := range seriesCounts {
-		b.Run(fmt.Sprintf("%d series", seriesCount), func(b *testing.B) {
+		b.Run(fmt.Sprintf("series=%d", seriesCount), func(b *testing.B) {
 			for _, samplesPerAppend := range []int64{1, 2, 5, 100} {
-				b.Run(fmt.Sprintf("%d samples per append", samplesPerAppend), func(b *testing.B) {
+				b.Run(fmt.Sprintf("samples_per_append=%d", samplesPerAppend), func(b *testing.B) {
 					h, _ := newTestHead(b, 10000, wlog.CompressionNone, false)
 					b.Cleanup(func() { require.NoError(b, h.Close()) })
 
@@ -161,6 +161,52 @@ func BenchmarkHeadAppender_Append_Commit_ExistingSeries(b *testing.B) {
 							var ref storage.SeriesRef
 							for sampleIndex := int64(0); sampleIndex < samplesPerAppend; sampleIndex++ {
 								ref, err = app.Append(ref, s.Labels(), ts+sampleIndex, float64(ts+sampleIndex))
+								if err != nil {
+									return err
+								}
+							}
+						}
+						ts += 1000 // should increment more than highest samplesPerAppend
+						return app.Commit()
+					}
+
+					// Init series, that's not what we're benchmarking here.
+					require.NoError(b, appendSamples())
+
+					b.ReportAllocs()
+					b.ResetTimer()
+
+					for i := 0; i < b.N; i++ {
+						require.NoError(b, appendSamples())
+					}
+				})
+			}
+		})
+	}
+}
+
+func BenchmarkHeadAppender_Append_Commit_ExistingSeries_HistogramValues(b *testing.B) {
+	seriesCounts := []int{100, 1000, 10000}
+	series := genHistogramSeries(10000, 10, 0, 0, 1, false)
+
+	numHistograms := 1 << 11
+	histograms := tsdbutil.GenerateTestHistograms(numHistograms)
+	for _, seriesCount := range seriesCounts {
+		b.Run(fmt.Sprintf("series=%d", seriesCount), func(b *testing.B) {
+			for _, samplesPerAppend := range []int64{1, 2, 5, 100} {
+				b.Run(fmt.Sprintf("samples_per_append=%d", samplesPerAppend), func(b *testing.B) {
+					h, _ := newTestHead(b, 10000, wlog.CompressionNone, false)
+					b.Cleanup(func() { require.NoError(b, h.Close()) })
+
+					ts := int64(1000)
+					appendSamples := func() error {
+						var err error
+						app := h.Appender(context.Background())
+						for si, s := range series[:seriesCount] {
+							var ref storage.SeriesRef
+							for sampleIndex := int64(0); sampleIndex < samplesPerAppend; sampleIndex++ {
+								histogram := histograms[(int(sampleIndex)+si)%numHistograms]
+								ref, err = app.AppendHistogram(ref, s.Labels(), ts+sampleIndex, histogram, nil)
 								if err != nil {
 									return err
 								}