@colega 2024-07-31
I originally used `uintptr` here, but wasn't sure whether that was properly handled by GC, and saw that it was actually unneeded, we can just use `unsafe.Pointer` which is _safer_, and more importantly: more correct, because it's actually incorrect to convert `uintptr` to `unsafe.Pointer`, from the `unsafe.Pointer`'s docs:

```
// The following patterns involving Pointer are valid.
// ...
// (2) Conversion of a Pointer to a uintptr (but not back to Pointer).
//
// Converting a Pointer to a uintptr produces the memory address of the value
// pointed at, as an integer. The usual use for such a uintptr is to print it.
//
// Conversion of a uintptr back to Pointer is not valid in general.
```