@colega 2024-08-01
I ran the benchmarks again with `-tags stringlabels -cpu 1` and they look more consistent now:
```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
                                                                                              │ main-1cpu-stringlabels │        new-1cpu-stringlabels        │
                                                                                              │         sec/op         │   sec/op     vs base                │
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=100/samples_per_append=1                      114.9µ ± 2%   104.5µ ± 2%   -9.06% (p=0.000 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=100/samples_per_append=2                      271.8µ ± 5%   238.3µ ± 1%  -12.33% (p=0.000 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=100/samples_per_append=5                      704.3µ ± 5%   655.4µ ± 6%   -6.95% (p=0.000 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=100/samples_per_append=100                    4.199m ± 6%   4.410m ± 2%        ~ (p=0.353 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=1000/samples_per_append=1                     2.188m ± 6%   2.167m ± 8%        ~ (p=0.481 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=1000/samples_per_append=2                     4.499m ± 4%   4.365m ± 3%        ~ (p=0.123 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=1000/samples_per_append=5                     4.474m ± 6%   4.457m ± 2%        ~ (p=0.684 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=1000/samples_per_append=100                   9.260m ± 2%   9.163m ± 1%        ~ (p=0.105 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=10000/samples_per_append=1                    4.541m ± 5%   4.464m ± 5%        ~ (p=0.165 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=10000/samples_per_append=2                    5.928m ± 3%   5.990m ± 4%        ~ (p=0.280 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=10000/samples_per_append=5                    8.018m ± 2%   8.098m ± 6%        ~ (p=0.105 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=10000/samples_per_append=100                  61.95m ± 5%   62.70m ± 2%        ~ (p=0.280 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=100/samples_per_append=1                  570.4µ ± 6%   595.5µ ± 1%   +4.41% (p=0.003 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=100/samples_per_append=2                  1.487m ± 3%   1.476m ± 2%        ~ (p=0.353 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=100/samples_per_append=5                  4.576m ± 6%   4.444m ± 3%        ~ (p=0.105 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=100/samples_per_append=100                6.145m ± 3%   6.127m ± 2%        ~ (p=0.481 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=1000/samples_per_append=1                 4.505m ± 4%   4.480m ± 5%        ~ (p=0.796 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=1000/samples_per_append=2                 4.193m ± 4%   4.196m ± 3%        ~ (p=0.853 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=1000/samples_per_append=5                 5.028m ± 5%   4.945m ± 3%        ~ (p=0.218 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=1000/samples_per_append=100               36.09m ± 3%   35.51m ± 2%        ~ (p=0.143 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=10000/samples_per_append=1                8.432m ± 3%   8.183m ± 2%   -2.95% (p=0.002 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=10000/samples_per_append=2                20.13m ± 1%   20.16m ± 2%        ~ (p=0.796 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=10000/samples_per_append=5                30.97m ± 3%   30.64m ± 2%        ~ (p=0.436 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=10000/samples_per_append=100              332.1m ± 4%   333.5m ± 5%        ~ (p=0.684 n=10)
geomean                                                                                                    5.039m        4.965m        -1.48%

                                                                                              │ main-1cpu-stringlabels │         new-1cpu-stringlabels         │
                                                                                              │          B/op          │     B/op      vs base                 │
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=100/samples_per_append=1                     3.931Ki ± 0%   3.931Ki ± 0%       ~ (p=1.000 n=10) ¹
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=100/samples_per_append=2                     3.967Ki ± 0%   3.967Ki ± 0%       ~ (p=1.000 n=10) ¹
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=100/samples_per_append=5                     5.096Ki ± 0%   5.096Ki ± 0%       ~ (p=1.000 n=10) ¹
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=100/samples_per_append=100                   62.16Ki ± 0%   62.17Ki ± 0%       ~ (p=0.109 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=1000/samples_per_append=1                    36.64Ki ± 0%   36.61Ki ± 0%       ~ (p=0.171 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=1000/samples_per_append=2                    37.03Ki ± 0%   37.01Ki ± 0%       ~ (p=0.618 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=1000/samples_per_append=5                    61.80Ki ± 0%   61.80Ki ± 0%       ~ (p=0.591 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=1000/samples_per_append=100                  610.4Ki ± 0%   610.4Ki ± 0%  +0.00% (p=0.020 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=10000/samples_per_append=1                   355.8Ki ± 1%   360.4Ki ± 2%  +1.31% (p=0.049 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=10000/samples_per_append=2                   356.1Ki ± 2%   348.5Ki ± 2%       ~ (p=0.137 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=10000/samples_per_append=5                   582.1Ki ± 2%   574.2Ki ± 2%  -1.37% (p=0.014 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=10000/samples_per_append=100                 5.952Mi ± 0%   5.952Mi ± 0%       ~ (p=0.270 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=100/samples_per_append=1                 12.40Ki ± 0%   12.40Ki ± 0%       ~ (p=1.000 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=100/samples_per_append=2                 103.4Ki ± 0%   103.4Ki ± 0%       ~ (p=0.340 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=100/samples_per_append=5                 108.7Ki ± 0%   108.7Ki ± 0%       ~ (p=0.403 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=100/samples_per_append=100               579.3Ki ± 0%   579.3Ki ± 0%       ~ (p=0.641 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=1000/samples_per_append=1                116.5Ki ± 0%   116.7Ki ± 0%       ~ (p=0.436 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=1000/samples_per_append=2                1.008Mi ± 0%   1.008Mi ± 0%       ~ (p=0.899 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=1000/samples_per_append=5                1.110Mi ± 0%   1.110Mi ± 0%       ~ (p=0.448 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=1000/samples_per_append=100              6.850Mi ± 0%   6.850Mi ± 0%       ~ (p=0.540 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=10000/samples_per_append=1               1.248Mi ± 2%   1.238Mi ± 1%       ~ (p=0.468 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=10000/samples_per_append=2               10.34Mi ± 0%   10.34Mi ± 0%       ~ (p=0.782 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=10000/samples_per_append=5               11.69Mi ± 0%   11.69Mi ± 0%       ~ (p=0.810 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=10000/samples_per_append=100             68.83Mi ± 0%   68.83Mi ± 0%       ~ (p=0.805 n=10)
geomean                                                                                                   287.4Ki        287.0Ki       -0.13%
¹ all samples are equal

                                                                                              │ main-1cpu-stringlabels │        new-1cpu-stringlabels         │
                                                                                              │       allocs/op        │  allocs/op   vs base                 │
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=100/samples_per_append=1                       51.00 ± 0%    51.00 ± 0%       ~ (p=1.000 n=10) ¹
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=100/samples_per_append=2                       51.00 ± 0%    51.00 ± 0%       ~ (p=1.000 n=10) ¹
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=100/samples_per_append=5                       56.00 ± 0%    56.00 ± 0%       ~ (p=1.000 n=10) ¹
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=100/samples_per_append=100                     609.0 ± 0%    609.0 ± 0%       ~ (p=1.000 n=10) ¹
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=1000/samples_per_append=1                      504.0 ± 0%    504.0 ± 0%       ~ (p=1.000 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=1000/samples_per_append=2                      509.0 ± 0%    508.5 ± 0%       ~ (p=1.000 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=1000/samples_per_append=5                      607.5 ± 0%    606.5 ± 0%       ~ (p=0.353 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=1000/samples_per_append=100                   6.009k ± 0%   6.009k ± 0%       ~ (p=1.000 n=10) ¹
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=10000/samples_per_append=1                    4.944k ± 1%   5.009k ± 2%       ~ (p=0.053 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=10000/samples_per_append=2                    4.949k ± 2%   4.843k ± 2%       ~ (p=0.117 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=10000/samples_per_append=5                    5.826k ± 2%   5.753k ± 2%       ~ (p=0.136 n=10)
HeadAppender_Append_Commit_ExistingSeries_FloatValues/series=10000/samples_per_append=100                  60.01k ± 0%   60.01k ± 0%       ~ (p=0.350 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=100/samples_per_append=1                   120.0 ± 0%    120.0 ± 0%       ~ (p=1.000 n=10) ¹
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=100/samples_per_append=2                  1.112k ± 0%   1.112k ± 0%       ~ (p=1.000 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=100/samples_per_append=5                  1.119k ± 0%   1.119k ± 0%       ~ (p=1.000 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=100/samples_per_append=100                1.437k ± 0%   1.437k ± 0%       ~ (p=1.000 n=10) ¹
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=1000/samples_per_append=1                 1.118k ± 0%   1.119k ± 0%       ~ (p=0.424 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=1000/samples_per_append=2                 11.02k ± 0%   11.02k ± 0%       ~ (p=1.000 n=10) ¹
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=1000/samples_per_append=5                 11.03k ± 0%   11.03k ± 0%       ~ (p=1.000 n=10) ¹
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=1000/samples_per_append=100               14.05k ± 0%   14.05k ± 0%       ~ (p=1.000 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=10000/samples_per_append=1                10.91k ± 3%   10.80k ± 1%       ~ (p=0.466 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=10000/samples_per_append=2                110.0k ± 0%   110.0k ± 0%       ~ (p=1.000 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=10000/samples_per_append=5                110.1k ± 0%   110.1k ± 0%       ~ (p=0.653 n=10)
HeadAppender_Append_Commit_ExistingSeries_HistogramValues/series=10000/samples_per_append=100              141.6k ± 0%   141.6k ± 0%       ~ (p=0.656 n=10)
geomean                                                                                                    2.545k        2.541k       -0.14%
¹ all samples are equal
```